SONARJAVA-6759: Extract shared NaN detection helper and improve test coverage
Move duplicated NaN detection logic from FloatEqualityCheck and
NanEqualityCheck into ExpressionUtils.getNanOwnerTypeName(), and add
additional test cases for Float.NaN parenthesized and self-comparison
patterns to improve coverage on new code.
